% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_responses_array.R
\name{code_responses_array}
\alias{code_responses_array}
\title{Automatically code responses with coding scheme}
\usage{
code_responses_array(coding_scheme, responses)
}
\arguments{
\item{coding_scheme}{List. Coding scheme of a unit from IQB Studio Lite.}

\item{responses}{List. Unit responses from IQB Testcenter.}
}
\value{
A data frame. Please not that the value column is coerced to a character, i.e., list entries of `c("a", "b")` are concatenated to `[[a,b]]`.
}
\description{
Automatically code responses with coding scheme
}
