% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_unit_definition.R
\name{extract_unit_definition}
\alias{extract_unit_definition}
\title{Automatically code responses with coding scheme}
\usage{
extract_unit_definition(unit_definition)
}
\arguments{
\item{unit_definition}{List or character. Unit definition of a unit from IQB Studio.}
}
\value{
A data frame.
}
\description{
Automatically code responses with coding scheme
}
