% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependency_tree.R
\name{get_dependency_tree}
\alias{get_dependency_tree}
\title{Displays the variable dependency tree of a given coding scheme}
\usage{
get_dependency_tree(coding_scheme)
}
\arguments{
\item{coding_scheme}{List. Coding scheme of a unit from IQB Studio Lite.}
}
\value{
A data frame.
}
\description{
Displays the variable dependency tree of a given coding scheme
}
