% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_responses.R
\name{code_responses}
\alias{code_responses}
\title{Automatically code responses with coding scheme}
\usage{
code_responses(coding_scheme, responses)
}
\arguments{
\item{coding_scheme}{List. Coding scheme of a unit from IQB Studio Lite.}

\item{responses}{List. Unit responses from IQB Testcenter.}
}
\value{
A list.
}
\description{
Automatically code responses with coding scheme
}
