% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variable_location.R
\name{extract_variable_location}
\alias{extract_variable_location}
\title{Extract variable locations from unit definition}
\usage{
extract_variable_location(units)
}
\arguments{
\item{units}{Tibble. Must contain a column `definition` that contains unit definitions. Please note, that it is recommended to not extract more than 20 unit definitions at a time as the call is memory-consuming.}
}
\value{
A data frame.
}
\description{
Extract variable locations from unit definition
}
