% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_unit_definition.R
\name{extract_variable_location}
\alias{extract_variable_location}
\title{Extract variable locations from unit definition}
\usage{
extract_variable_location(units)
}
\arguments{
\item{units}{Tibble. Must contain a column `definition` that contains unit definitions. The column}
}
\value{
A data frame.
}
\description{
Extract variable locations from unit definition
}
